refactor: KidError to submodule jwk
